AthAllocators: Fix memory leak when context argument is supplied in serial job.
There are two versions of ArenaHeader::allocator, one taking an EventContext
argument and one not.  The one without a context was first checking
for an arena set via the thread-specific pointer m_arena, and if that's
not set, using m_defaultArena.

The version that takes an EventContext was first trying to look up an
arena using the event slot, and if that fails, using m_defaultArena.

The problem comes if you supply the EventContext argument in a non-MT job.
Since it's not MT, we won't have slots set up, so we use m_defaultArena.
However, this is usually not correct --- we need to use the version
set through the thread-specific pointer, which will be pointing at the
arena in the event store.  The default arena will never be cleared
by default --- so things will appear to work, but you'll have a big memory leak.

Fixed logic to properly check the thread-specific pointer if the lookup
by slot doesn't work.


Former-commit-id: 49a52825d77c850ff6533314e302e992bbde4308