CaloRec: Preparing to make CaloCellFastCopyTool const.
Remove writes to member variables from process().
This was used to cache the decision about which strategy to use to do the copy.
However, this is only a couple logical tests to begin with, so hardly anything
to be gained by caching it, and the cache spoils constness.

Make all internal methods called by process() const.


Former-commit-id: cc2e1fd7e23cb4e7737ad5e9931c42a8dca63fd2