AthenaBaseComps: Make sure to start() conditions handle keys recursively.
In the AthAlg base classes, we extend start so that it calls a start()
method on all registered read conditions handle keys.  This is to allow
them to be able to find and cache a pointer to the conditions
container (initialize() is often too early for this).
However, we were not doing this for private tools owned by the algorithms.
Extended so that we visit all the tools recursively.



Former-commit-id: 1c511d6e0924f48f500c324007139577bcfed92e