Minor addition to tau trigger online monitoring
This MR adds a few variables (RNN tau ID) to the tau trigger online monitoring.
A few monitored variables were also dropped because unused.
It would be nice to have this in the release to be deployed for post-TS1 high-mu data.
This is why I set the 'urgent' label, but it's not actually urgent.

Thanks.
Bertrand


Former-commit-id: 6a13497119a54bc909ff8ac03399ac6d8048d08a