Ending writing of transient decorations on Muons in DRAW_ZMUMU building
ATLASRECTS-4499

The previous version of the DRAW_ZMUMU code used derivation framework tools to decorate muons with a "DFCommon" flag for the purposes of event selection. Unfortunately thesevariables were getting pulled into the AOD (and other formats) by the ItemList, and this was causing downstream problems.

A temporary fix was expedited by editing the ItemList and manually removing these decorations, but a better solution is to run the muon selector tool directly from a new AlgTool which is provided here, rather than going via decorations. This change now introduces the tools and allows the decorations to be dropped.


Former-commit-id: be90145f9898f89df84f2cd7dcb07e505020be0d