Not defining unknown PDG IDs for ATLASSIM-3007
There are a few PDG IDs that were being defined in the RHadrons code
that were *standard model* particles that could in principle come from
an RHadron decay.  It's likely these particles weren't quite being
treated properly in the new setup, so we've removed their definition.
Now if they appear, we should get a crash during simulation (unknown
particle), and we can look into the job and what exactly is happening
(both why they are produced and how they are interacting).


Former-commit-id: ee3be3787a9fdf3602f6a15b5e4b44f089ac2776