Move CreateFakeStep method into the SimulationEnergies class.
Add a new `CreateFakeStep` method to the `SimulationEnergies`
class based on the version in `(Tile)EscapedEnergyProcessing`,
but incorporating some of the functionality from the
`SimulationEnergies::ProcessEscapedEnergy` method. This is then called
from `SimulationEnergies::ProcessEscapedEnergy` prior to the
calls to `(Tile)EscapedEnergyProcessing::Process`. Remove
obsolete `CreateFakeStep` methods from
`(Tile)EscapedEnergyProcessing` classes.

Drop old `Process` method and rename `ProcessNew` to `Process` in
`(Tile)EscapedEnergyProcessing`. Switch `ProcessEscapedEnergy`
to provide a `G4Step` as input to the `Process` calls.

Part of the tidying for ATLASSIM-3718.


Former-commit-id: e2efc345963b0b26b20dcaa5e544934266bb7981