start FTagAnalysisAlgorithms package
First version of the package.  In practice it is only one algorithm
(as the selection tool is a standard selection tool), and a set of
example configuration files.  In the configuration I had to disable
one of the algorithms, because I got some sort of index error,
hopefully the experts can fix that.


Former-commit-id: 660c3517aca336fbf75cd49131a770febf5ed5cc