AthenaBaseComps: Work around thread-safety issue with Algorithm.
Algorithm::sysExecute takes an EventContext as an argument and saves
it as a member variable.  If the context has an extension block and this
is a reentrant algorithm running in MT mode then this can lead to a
double-delete of an extension block.  Work around by overriding sysExecute
in AthReentrantAlgorithm to clear the extension block before passing
it to the base class.  (Doesn't matter, since a reentrant algorithm will
never use this copy of the context anyway.)

Former-commit-id: 2e3d04d11062e07f76583331adff59126e002d91