Enable ALFA ROB monitoring by default.
As discussed in ATR-17748, since the ALFA ROB monitoring does not seem to cause any problems when enabled for standard physics data taking, we will enable it by default in future SMKs. This way it will not be forgotten in the event that there is some special runs with ALFA where this needs to be enabled.


Former-commit-id: cae626c5edede45a58dd9cf78c4161c286b5f1c1