L1TopoSimulation: Remove deprecated beginRun method
begin/endRun() are deprecated. Moved the resetting of the scaler to
start() instead. This is not entirely equivalent but for the use-cases
we have this should be good enough (for MC the run number never changes,
and most reco jobs only process one run). For online it will be execute
for each run number change.


Former-commit-id: 3cc31f87997d13b450acafb9b0b7e164d6b5a41d