AthenaServices: Get MakeInputDataHeader working again.
The MakeInputDataHeader algorithm is intended to mark an output DH
as an input one, so that it will be included in another output DH
for back-navigation.  However, it wasn't MT compatible, because
it was directly changing the header in SG.  Change so that we can
mark a header as input by adding an alias with a key ending
in _Input.



Former-commit-id: 6ee1949ebe4cde1d24a210030eab6fe8229a9bcc