Fixes to get BS->RDO decoding working correctly for L2MuonSA
In RpcDataPreparator, added the BS->RDO decoding for the RoI driven
mode (previously we had added the RS->RDO decoding for the full dector,
but missed the RoI driven case).

In mu.withViews and muMenuDefs I made sure sure the L2 RpcDataPreparator is
properly configured for the BS->RDO decoding. I also removed the
muonRecFlags.doCSCs=True line, which was causing the
DetFlags.readRDOBS.RPC_on flag to be set to False
(which is incorrect for data running). I checked that
the CSC flag is already True, so that line is not needed.


Former-commit-id: 707ac241fb8e4f47668fb56acbd092d90fc21e06