Update MdtOverlay, RpcOverlay and TgcOverlay for Multi-threading
This commit updates `MdtOverlay`, `RpcOverlay` and `TgcOverlay` to use
`SG::Read/WriteHandleKeys` to specify their main inputs and output (DigitContainers)
It also separates the creation of the two input DigitContainers (from RDOs and HITS)
into separate Algorithms to make the data-dependencies clearer.

The functionality to be able to copy the hard-scatter DigitContainers into a separate
StoreGate has been removed as this is not used currently. It will be simple to add
this back at a later date if required.

Some direct StoreGate access occurs when hard-scatter SDOs are copied to the output
StoreGate. This can easily be recified by updating the corresponding base-class methods.


Former-commit-id: 735a738666df28bb8fe69b0b0401ad8bc8cf1f1d