Pulling in updates for FastSiDigitization package from 21.3
This merge request pulls in update to the FastSiDigitization package
from the ~"21.3" branch, so that the code matches, other than in the
case of known migrations that have taken place in ~master (e.g. the
migration to Conditions Algorithms in the SCT.)

Changes related to the `HepMcParticleLink` migration have been left
out as they are not yet ready to go into ~master.


Former-commit-id: f739c72f450c277e8cfff9268ed26c2b36157f41