Patch for ATLJETMET-973, bad particles entering truth WZ jets
This patch solves two problems with the selection of particles for the
truth WZ jets.  "Standard" truth jets are unaffected.  The WZ selection
is meant to exclude true prompt electrons and muons from the jets.  The
selection of true prompt electrons and muons accidentally did not
include a check that the leptons were generator stable (status and
barcode), and so upstream particles were being included.  This was the
origin of the report that "unknown" leptons were entering the truth
jets.  While this didn't affect the truth jets in most cases, the
leptons are also used for dressing, so photons were sometimes also being
removed from the truth jets.

The second problem affected neutrinos.  The neutrino usage was altered
slightly, leaving the default flag configuration in good shape, but also
leading to a missed case.  When leptons are meant to be excluded (as in
the case of WZ truth jets), prompt neutrinos should also be excluded.
There was no check for prompt neutrinos, so they were still being
included in the truth jets.  This lead to a significant excess of truth
jets (coming from the neutrinos).


Former-commit-id: f29cb5b3a4df0fdfcae4ce342fa0928e9618c197