Add "HardScatterSplittingMode" and "ParticleBarcodeVeto" properties to SiSmearDigitizationTool
All other InnerDetector PileUpTools have the `HardScatterSplittingMode` and `ParticleBarcodeVeto` properties,
so have added these to `SiSmearDigitizationTool` to ensure consistent behaviour.

The default behaviour is still to digitize all bunch-crossings sent to the tool, but setting `HardScatterSplittingMode`
to either 1 or 2 will cause the PileUpTool to either process only the hard-scatter SiHits or only the pile-up SiHits.

The `ParticleBarcodeVeto` property prevents PRD Truth objects being created for particles associated to Geantinos
in the truth record. Geantinos are used to flag that SiHits are associated with pile-up events. Not creating SDOs or
PRD Truth for such hits is the default behaviour for all other ID PileUpTools.


Former-commit-id: 4de37dc18df5b409a56e09e0a4f6cbd260be9968