AthenaCommon: Work around bug in python 2.7 readline module.
The python readline module as of 2.7 has a bug:

errno = write_history(s);
if (!errno && _history_length >= 0)
    history_truncate_file(s, _history_length);
if (errno)
    return PyErr_SetFromErrno(PyExc_IOError);

History file truncation is controlled by _history_length, which is
set by set_history_length; athena usually sets it to 1024.
If this has been set then history_truncate_file is called.
If the history file has more lines than this setting, then the
file will be truncated.  This internally calls readlink()
on the history file.  In the usual case that this isn't
a symlink, this fails and sets errno to EINVAL.  readline handles
this properly and returns successfully.  The readline module,
however, sees at this point that errno is set and returns an
error to python.  The upshot is that we get an annoying error
on exit.  Suppress this by using a wrapper around the
write_history_file function that catches and suppresses the exception.


Former-commit-id: 47357fa8c33480d8228a9e59e212e04ef374798b