fix listing of CP tools (ANALYSISTO-627)
top::listToolStore() errorneously skipped tools with a dot in their name
when parsing the output from asg::ToolStore::dumpToolConfig(). As a
result, the overlap removal tools were not listed.

Relax the regular expression for matching the tool name such that all
names are accepted as long as they don't contain whitespace.


Former-commit-id: 90c43a4a7bc5760208c37bb2d45d9cdf606fc321