Made algorithm sequences more robust against name clashes.
On Nils's suggestion re-designed a bit how output container names are constructed for the algorithms. Basing the container names on the output names by default. And allowing the user to specify a prefix for any "hidden layers" of objects that a really complicated algorithm sequence may have.

At the same time fixed the logic with which the affecting systematics are propagated for multi-input / multi-output algorithms. The previous logic was appending regular expressions way too liberally.


Former-commit-id: 5ada2f27a853475b487076c07ae7e1d493b93930