Taught CP::AsgViewFromSelectionAlg how to perform deep copies.
Instead of creating a separate algorithm for it, it seemed easier to just extend this existing algorithm for creating deep copies of particle containers.
At the same time tweaked the logic in the view copy making code a bit as well. So that immutable input containers could be used by the algorithm as well.


Former-commit-id: 5cbe3969fa7d61a6d5446620a24b844bfb1ca6cf