Fix for ATLASRECTS-4510
The MuonTrackSummaryHelperTool produces multiple warnings for muons in the case that the number of hits+holes for a given CSC, TGC, or RPC chaber does not match the number of gas gaps in that chamber.
This can occur because the final cleaning stage, in which outliers are removed from tracks, occurs after the final hole-finding stage.
Therefore there can be missing CSC, TGC, or RPC holes, leading to this warning.  The solution is simply to have the cleaner add a hole when it removes a CSC, TGC, or RPC outlier.
This will not affect analysis working points because:
1. No hits are removed by this fix: the hit is already removed by the cleaner.
2. The number of precision hole layers is only used for muons with |eta|<0.1, hence CSC holes cannot contribute here.
3. The only other place where hole counts are used in the selector tool is the high-pT WP, which uses the number of extended small holes: this does not include any CSC, TGC, or RPC holes.
As a result, although this is an FT0-violating change it will be totally transparent to analyzers.
No significant changes were seen when running on the usual muon physics validation samples: https://indico.cern.ch/event/748044/contributions/3095388/attachments/1696707/2731205/muon_validation_020818.pdf
This fix was discussed in the reconstruction meeting of August 14, see https://indico.cern.ch/event/750102/contributions/3106058/attachments/1700609/2738628/ATLASRECTS-4510.pdf


Former-commit-id: 2bf4190f9e9e31134ace7177815269292e873269