ATLASRECTS-4540: Add protection for invalidated vector iterators in SiTrackMaker_xk
Recentely, the std::list objects in many tracking packages were changed to
std::vectors.  Unfortunately, the behaviour between the two is not exactly
identical, as erasing an element in std::list will never invalidate the
iterator, while doing the same to a vector will.  Following the discussion
in ATLASRECTS-4540 some protection for this was added to prevent crashes.

Former-commit-id: f1419e80c07aaee2095a50294c2aee1d0da8b8b7