Merged the executeTemplate and executeDeepTemplate functions.
In the end, while it having them as two separate functions could probably result in an infinitesimally faster binary, code maintenance-wise it was not a good choice to have two functions with such a large overlap. So now the same function takes care of making both shallow and deep copies.


Former-commit-id: 527807eae1a31033b5f134de79d3946d62805488