Basic resolution support
Added support for JER, JPtR, and JMR in order to prepare for
recommendations.  This is not yet complete, as the nominal
histograms are not yet read from files.  Furthermore this is
not yet validated.  This is an intermediate commit where key
functionality has been added and debugged until it compiles,
but not yet fully connected with the configuration reading
stages and related user input portions.


Former-commit-id: 993e68769c5a66cbe7a954fd5c0b0753d7def9e6