Taught the standalone AlgSequence how to deal with sub-sequences.
In order to be able to construct a bit more elaborate algorithm
sequences using exactly the same code for Athena and EventLoop, taught
the standalone version of AlgSequence how to swallow and then handle
AlgSequence objects added to it.

Added a bunch of unit tests to the code at the same time.


Former-commit-id: c57d4e0efd730663ecffaa8bf574e1b5e9b7e7c2