Refactoring the ISF_Fatras Configuration
This commit moves the `CfgGetter` methods for various `ISF_Fatras`
tools into the packages which contain the C++ implementation of the
tools. This is inline with how the configuration for other ISF code
is done.

`Cfgetter` methods for `Tracking` Tools and Services are moved into
a separate module within the `ISF_Services` package.

At the same time xplicit calls to `getService` and `getPublicTool`
are dropped.


Former-commit-id: f39c5abcaba0e7543d5e259acdbac2f0b6683fb1