Added one-dimensional interpolation for resolution
Following discussions with the people tagged on the MR, it was pointed
out that the 2D histograms of (pT, eta) should be interpolated across
pT but not across eta.  This MR adds the needed functionality to
support this behaviour in a configurable way, such that JMR can
continue to be interpolated across both dimensions.  This also adds
support to both JetUncertainties and JetCalibTools.


Former-commit-id: a758d993e15c7233c4f1ecac401e058cd57e3173