Fixed the code's logic when cutting on the cluster eta value.
The code was not taking into account that the casting to xAOD::Electron (now xAOD::Egamma)
was made part of the TAccept cutflow. And hence cuts on cluster eta values would never
result in anything being accepted.

At the same time stopped using xAOD::Electron in the code, switching to xAOD::Egamma instead.
To allow making such cuts on photons as well.


Former-commit-id: 1818e6977ce09f5edd53293a1fe6fc9fea6f1505