Fixed up the electron analysis sequence to behave correctly/better for data input.
Stopped running the efficiency scale factor calculation when running on data input, as it would always result in a 'bad_eff' result.
Changed the code to always perform the electron likelihood selection at the beginning of the sequence as it's the recommendation, and is faster than doing it later on.
Had to re-organise a bit the way the selection code is set up.


Former-commit-id: a2bf2b479cc1db93837238c94597ba0aff584dfe