Protect SimKernelMT_test in ISF_Algorithms against dynamic_cast failures
Technically it is possible that the `dynamic_cast` in `SimKernelMT_test` could fail
and a nullptr could be dereferenced. The complication is that `ASSERT_NE` is expanded
into code that returns `void` whereas `EXPECT_NE` does not. Therefore `ASSERT_NE` cannot
be called from within functions which do not return `void`. These changes should fix
Coverity Defects 118535 and 118536 in ATLASSIM-3686.


Former-commit-id: 76c36d93693426a7dbcec79c1e49582a622d8915