Added missing sanity checks when setting KLFitter likelihood
These checks are necessary as setting the likelihood for KLFitter might fail. If
not checked, this could run into seg faults when trying to perform the fitting.
In particular with a pending update of the KLFitter version, checking this
return code will be important to make sure that the KLFitter configuration is
correct.


Former-commit-id: d397627b6713f2cf1415517a217abcdc91b04851