A few developer-only bugfixes and minor tweaks
- Fixed a developer bug where testing files specified with full
paths were sometimes incorrectly read in (leading to the tool
failing to initize).  This was never supported behaviour for users,
so no users should be affected, and if the were then the tool
would have failed to initialize.

- Fixed a developer bug in the validation script where the
new resolution IsData parameter was not being set.  This is now
set and thus the validation script works for new resolution
uncertainties.

- Added a print message for the IsData parameter to the normal
initialization info dump.  This changes nothing, but it will make
it easier to debug user problems in the future, as all of the
parameters that they used to configure the tool are in stdout.

- Updated the UnknownFlavourComp.root file to include PFlow jets,
like it does in CalibArea.  This is really only for backward
compatibility, as users are recommended to use CalibArea.  This is
the default flavour uncertainties "50/50" file.


Former-commit-id: bbaadb4acbd7d8b6ff186d19d9914ea5701ea225