Apply Parallel-running on ART Test for MuonRecRTT
ART Parallel-running can shorten the running time of MuonRecRTT test.
Hopefully, this mechanism can also decrease the requirement of disk
space. We don't want to use art-split because that will lead to too
many subjobs.
At the same time, the number of input files is reduced from 10 to 6
for each job, to avoid failure caused by insufficient disk space.
In addition, for master branch, the validation test is cancelled
because there is some error that will cause the jobs to fail.


Former-commit-id: 5863866a0c5c9516918e751e481e0355ad96e144