Nominal resolution retrieval in JetCalibTools
It is now possible to retrieve the nominal MC and data jet resolution
from JetCalibTools.  This is functionality needed for the calculation
of object-based MET significance.  This extends the IJetCalibrationTool
interface to include this functionality.

The JetCalibTools resolution validation script was also extended in
order to retrieve the nominal resolution, showing that it is
working as expected.

In parallel, there are a few small changes to the JetUncertainties
plotting scripts, made for the JER recommendation plots.  No user
code has been changed in JetUncertainties.


Former-commit-id: c9697bc5346bada22627def30aece9237c858c48