Add an option of calling JpsiFinder in Reco_mumu tool
A job option to check collections before calling the JpsiFinder is implemented in the Reco_mumu tool. By default this flag is turned off. By setting the flag to be true, the Reco_mumu tool will check whether the collections provided by user are empty of not before calling the JpsiFinder. For example in our analysis of B_c+>J/psi D_s+, with such a flay in BPHY15 format, we could run the instance for di-track (which is used to form D_s+ candidate) only if a di-muon was already found. Hence the running speed is significantly increased.


Former-commit-id: 360c60d5609dfb8d3214e78758929eade70b2fe5