Updated the muon test jobs, so the Athena one would be able to run.
Using a wildcard for selecting the containers to write is not a good idea anymore,
since the AnaAlgSequence configuration bases the intermediate container names of
the sequences on the output container name.

For now just switched the jobs to only writing the systematically not varied objects
to the mini-xAOD output.


Former-commit-id: 9ddeb787a93a7a8cf4718428180b7c73d7f183f2