Update Muon EF MS only hypo for multi muon hypo
This extends the functionality of the hypo to be able to handle multi-muon chains, similar to the other muon hypos. It also removes the MSonly hypo from TrigMuonHypoConfig, as for athenaMT we are now using TrigMuonHypoConfigMT and testTrigMuonHypoConfig.
This is essentially an update of https://gitlab.cern.ch/atlas/athena/merge_requests/13648 which was reverted because the configuration was not correct and causing tests to fail (the configuration is now correct here).


Former-commit-id: 7827ec02f52e79d45bb8f5dd029d0a31e8718ecb