SCT_Digitization: Fix ubsan warnings.
In gcc7, ubsan will warn if the distance between a vtable pointer
and the beginning of an object is greater than a sanity check threshold,
hardcoded as 1<<20.  This implies that classes larger than this that
use virtual derivation can get warnings from ubsan.       
While this is in principle a false positive, it's probably not
really a good idea to have such large class objects.  (gdb also  
complains, because in some cases it needs to copy objects.)

One such class is SCT_DetailedSurfaceChargesGenerator, which includes
a very large array.  This array is, however, not actually used,
so comment it out.  (Deliberately not removing it entirely at this point.)



Former-commit-id: 84227122046c13e76607980a54abbf5b30351001