Bug fix for HI jet aliasing (effectlvely sweep of !13933)
This merge included, for b-tagging purposes, the aliasing of
AntiKt4HI jet collection to AntiKt4EMTopo, which (at least
in a 21.0 test) was not enough for the EM Topo jet reference
data to be recognized. Aliasing is now extended to
AntiKt4TopoEM and AntiKt4LCTopo, to solve this issue and be
fully consistent with the scheme used for AntiKt4EMTopo.


Former-commit-id: a044d735452a78a7ce1ac5e75264be593a10faae