Bug fix for !12980: further aliasing added for AntiKt4HIJets
Previous merge for b-tagging support in heavy ion jets (!12980) included
the aliasing of AntiKt4HI jet collection to AntiKt4EMTopo, which was not enough
(at least in the case of a test with 21.0) for the EM Topo jet reference data
to be recognized. Aliasing is now extended to AntiKt4TopoEM and AntiKt4LCTopo,
to solve this issue and be fully consistent with the scheme used for AntiKt4EMTopo.


Former-commit-id: ee0ae7a85c2794261c306c0a47cbfe2878e2cfc2