Bug fix for !13696: further aliasing added for AntiKt4HIJets
Previous merge included, for b-tagging purposes, the aliasing of
AntiKt4HI jet collection to AntiKt4EMTopo, which was not enough
(at least in the case of one test with 21.0) for the EM Topo jet
reference data to be recognized. Aliasing is now extended to
AntiKt4TopoEM and AntiKt4LCTopo, to solve this issue and be fully
consistent with the scheme used for AntiKt4EMTopo.


Former-commit-id: c7c2ebcac5cc9e2a995a31f74de27a6426764f27