Improvement to DerivationFramework::PrescaleTool to give reproducible results when running in MP/MT
ATLASG-1450

It was noted that the derivation framework PrescaleTool is not safe in a multi-process/thread environment due to its use of an event counter, such that the order of execution needs to be the same to get reproducible results. This MR changes the tool to use the event number rather than a counter, and since the event number is independent of the the order of processing, the outcome of the prescaler is now exactly reproducible and predictable.


Former-commit-id: 34bb2bdbc7f28e74a954c5952ce0a7efb2fafc9b