Fix for MuonGMCheck
In BMG chambers, which have cutouts, it is not possible to simply loop through all tubs from 1 to N_tubes.
MuonGMCheck was doing this, which creates errors when it tries to access tubes that are not there.
This fix tests to make sure that a tube in a BMG chamber exists before trying to access it.


Former-commit-id: 36ddda3c4ca8b5d458a290b9805fc07bac1b7840