Fix for MuonGMCheck
In BMG chambers, which have cutouts, it is not possible to simply loop through all MDT tubes from 1 to N_tubes, as the tubes in the cutouts do not exist.
MuonGMCheck was doing this, which creates errors when it tries to access tubes that are not there.
This fix tests to make sure that a tube in a BMG chamber exists before trying to access it.
As MuonGMCheck is not used in reconstruction, no FT0 violation is expected or observed.  Results of RunTier0Tests can be found at /afs/cern.ch/work/w/wleight/public/MuonSW/MuonGeomRTT_21/.


Former-commit-id: bc1e00445b6ce88f1b8651f9750e626d95f0b43a