AppMgr.py: Fix bug in backwards-compatibility code
Fix a (probably) harmless bug in some of the backwards-compatibility
code in AppMgr. There was a double negation (`not not`) and a comparison
to the wrong configurable type (Service instead of AlgTool).

In addition made the code handle lists of tools correct and added a unit
test for ToolSvc.__iadd__


Former-commit-id: 3a61baeeb4fe4ac6b6ba2950d9371e729442a50a