Close input file only when all events from it finished processing
Change EventSelectorAthenaPool so it keeps track of which file the events
come from and closes input files only when all events from it finished
processing. The events are tracked used Begin/EndProcessing incidents
fired by the EventLoopMgr. File ID is extracted from the Extended Event
Context. In case the Manager does not fire incidents (like in serial
Athena) the files are closed right away, as before.


Former-commit-id: b9f751da6de5d147a953a152b51390435949eaec