Optimizing disk space usage by Archive_tf (ATLASJT-383)
Given the limited amount of scratch space available to SCORE jobs
in the grid, it was needed to further optimize disk usage by
Archive_tf. Zipped inputs are now deleted as soon as they are
unpacked to a temporary directory, which in turn is deleted as
soon as its content is saved to the output file.


Former-commit-id: 9a197aae6f6f6c97775a74c929e45ed82f9289c7