MuonMDT_CnvTools: Workaround for using MdtRdoToCscPrepDataTool as a private tool.
MdtRdoToCscPrepDataTool can be called from multiple parent tools,
but it keeps event data in member variables.  When it is made private,
then this fails, since we now have separate instances.
Hack around this so that tests will work again.
However, this class needs to be redesigned to work correctly
in MT jobs.

See ATR-18757.



Former-commit-id: f26a0f9d93de7070c173d4bf4f84a2d0c3bc93da