Fix for ATLASRECTS-4421
Currently the MuonCreatorTool adds two separate groups of segments to muons.
The TrackSegmentAssociationTool checks for segments that overlap with combined or SA tracks and adds those to the muon as the segmentsOnTrack auxdata.
Additionally, the muonSegmentLinks auxdata is filled with the segments used to reconstruct the segment-tagged muon that was reconstructed from the same ID track as the muon (assuming there is one).
Clearly, for combined or SA muons the first set of segments is far more important, while for segment-tagged muons the second set is what matters (the first doesn't exist, of course).
This fix simplifies the situation by consolidating the two into one: the muonSegmentLinks auxdata is now filled with the appropriate set of segments depending on the muon author.
As the segmentsOnTrack auxdata is then superfluous, it is removed.
This does mean that it is no longer possible to identify the segments used to reconstruct a segment-tagged muon from the ID track used by a combined muon.
However, it's hard to imagine a scenario in which this information would be used.  (The fact that the combined muon was also reconstructed by MuTagIMO is preserved.)


Former-commit-id: 08abe9ce2068b4dfff1d5c13be4d77c5cbb2d3da