Made IAthSelectionTool const
The accept methods of the IAthSelectionTool are now const.

To achieve const-ness of AthTruthSelectionTool, the original cut flow
object is split into a cut list and a cut flow part. The selection
tool only contains the cut list. The cut flow object is now external.

The selection tool still has the two modes ALL and UNTIL_FAIL, but there
are now different methods for the two modes, rather than a switch. And
the decorator and monitoring tool now use the UNTIL_FAIL mode.

Also the surfaces are now created during initialise if needed rather
then upon the first use.

This addresses ATLASRECTS-4567.


Former-commit-id: 9ac3650354f18be991087ecd6c9fc2d5596d7e0e