TileEvent: Move TileDQstatus here; make recordable.
Most of the work done by TileBeamInfoProvider needs to be changed to
an algorithm to work properly in reentrant MT jobs.  As a first step,
break out the TileDQstatus structure and adjust it so that it can be
recorded in SG.  Also add a few other data items produced by
TileBeamInfoProvider.


Former-commit-id: 86934877eb2a1e0b98fd5f4236cfdd615fb26aba