TileRecUtils: Move TileDQstatus to TileEvent.
Most of the work done by TileBeamInfoProvider needs to be changed to
an algorithm to work properly in reentrant MT jobs.  As a first step,
break out the TileDQstatus structure and adjust it so that it can be
recorded in SG.  Also add a few other data items produced by
TileBeamInfoProvider.


Former-commit-id: f0eb722c45f830b85ca35c4baf40dec3fbe49cac