Use the Reco_V0Finder in the K_S0 selection
Instead of the JpsiFinder tool, the Reco_V0Finder (which is dedicated to the reconstruction of long-lived particles such as K_S0, Lambda) is adopted to select the K_S0 candidates. As a result, the combinatorial background is significanly reduced. With less input of K_S0 candidates, the running speed of the JpsiPlusDs1Cascade algorithm is improved.


Former-commit-id: 82acee8a02b1ecc2905ed7c90693ac3db0af8217