Empty Pool File Creation for MuonGeomRTT on master Branch
Some of the MuonGeomRTT ART test scripts fail when running on master branch.
To fix this problem, the following line is added to each test scripts:
art.py createpoolfile

In addition, previous tests show that toth "ATLAS-R3-2021-00-00-00" and
"ATLAS-R2-2015-03-04-00" are not allowed for MuonGeantinoSim, so the script
are modified to use "ATLAS-R2-2016-01-00-01" for all branches.


Former-commit-id: df42efcb4c65a8dd6f4babdfa4dad0ca43339211