SUSYTools: Allow unknown ISO WPs to fallback to known WPs for SF determination
With the release of new isolation WPs, not all SFs are available yet for electron
muon, or trigger.  This change allows the new isolation WPs to be studied, while
falling back to a different isolation WP for the scale factor determination, while
printing a warning. This is just meant for expert use.


Former-commit-id: 39cfbd1ecb673a4beaa0fac1e26daf7475091358