Fixed checking for existence of truth-related containers (ATLHI-215)
The condition of checking for existence of truth-related containers
(needed to identify data overlay MC samples as simulation)
was crashing, because sometimes it would try to compare the string
content against a NoneType variable.
Now, the variable type is checked beforehand to avoid crashes.

In addition, HION3.py is modified to add some triggers that were
used in 2017 5 TeV pp data-taking.


Former-commit-id: c0daa845adfa6d30a5158144566fe5aa2e8c6a0d