Fix bug whereby a histogram plotted twice would be modifed twice
If setting an operation on a histogram, because of the annoying directory ownership,
if the object is retireved again from the file if it has already been modified, root
seems to have cached the modifed object and returns that rather than the object
requested from the file again.

So here we SetDirectory(0) such that it the object is later modified the modified
version is properly deleted by the delete as it should be, rather than hanging around
in some spurious fashion.


Former-commit-id: a4e1eb52ce50bc94ce72418f2ce8e9e805aab024