Tau Triggers: fix HLT tau-vertex element link
This MR fixes the broken element link from the HLT tau to its vertex in data.
Something similar happened for tau tracks and tau seed jet, as discussed in ATR-12096.
These were fixed, but the vertex link was overlooked.
I set the 'urgent' label because we would like to record 18 data with the link fixed.

Thanks.
Bertrand


Former-commit-id: 566dc45a4428a273fd8ead97c9bf7e3e06756ea1