CollectionBase: In test code, make sure files are closed before reopening.
cherry-picking commit 6020aae778 [formerly 478f5ee363].
The test in this package was leaking ICollection objects.
This meant that we could try to open a ROOT file that we still had open.
With root 6.14, this can cause an infinite recursion.
Make sure to delete ICollectionObjects when we're done with them,
so that the files get closed.


Former-commit-id: 6ce378364a89d374f840155911c4a693de8ed4ba