Set the default sTGC SimHit converter to _p1
The HepMcParticleLink_p2 format is not fully supported in master. To sweep the sTGC Simhit class
from 21.3 to master, usage of specific HepMcParticleLink_p2 parameters is replaced by the _p1
equivalent or removed. Also the default converter is switched to the _p1 converter. The _p2
converter is not erased in case it is needed in the future.


Former-commit-id: 01ba05ff5b454cd7cf954744b817bd875dfaefc9