Fix Tile Overlay with premixed pileup (ATLASSIM-3853)
In Overlay job TileRawChannelContainer is used to check DQ status,
but in this case it is supposed that real data are used for overlay
and this container is always available.
In the case of MC this container is available but it is useless though it's harmless.
In the case of premixed data this container is not available at all.

Compilation warnings are fixed also.


Former-commit-id: 10032d031c43f6524d92dc3fa9e179458192ff19