Updating Pythia8 RHadron setup
Adding functionality to programmatically generate lists of RHadrons,
their masses (from a set of options), their hadronic interactions, and
so on.  Uses the setups that have previously been used with Pythia6.

Adding some comments and documentation into the hadronic interaction
code as well, and cleaning up a few cases of confusing indentation.


Former-commit-id: 91d5c041764f7bc3af042a456e253a32fdeaa02b