DFSUSY trigger update
This MR is a follow up of !14565, to add missing unprescaled triggers for formats using trigger skimming.
In addition, the trigger skimming and thinning strategy of SUSY3 was revised, dropping the wildcard that was capturing tons of single-lepton triggers that were not primary physics triggers.
A few tau performance triggers were added for trigger efficiency measurement.
This more aggressive thinning&skimming strategy leads to huge savings, 35% smaller DAOD for data!
For SUSY5, a simple cleanup was made, following the modifications made in !14162 (prescaled trigger moved to SUSY17, no longer used in SUSY5).


Former-commit-id: c8bfc5187eee8e01c5eacc058e3c52b70041b394