Try to make configuration for RPC data decoding using ComponentAccumulator.
New file MuonRpcDataDecodeConfig.py tries to define a configuration
function that setups up the RPC raw data decoding algorithm and all
needed tools and services. It has a main function to allow testing.
Currently the the pickle file is created ok, but running the job
results in crashes when trying to access the RPC cabling DB.


Former-commit-id: a6b9342f83281abeb3c74dbc8fafe6632363ea87