Remove unnecessary set of phi maxima from MuonLayerHoughTool::Road
The Road stores the phi maxima associated to it in a vector.  But it also stores them in a std::set.
This set is only used once: prior to maxima sorting, the vector is emptied and then re-filled with the contents of the set.
But sets are ordered by definition: since the set is filled with pointers, using the set contents amounts to applying a pointer sorting to the phi maxima.
If no further sorting is then performed, the resulting set of maxima has been sorted by pointers, which makes irreproducibilities possible.
This fix removes that possibility by removing the set, which serves no purpose anyway.


Former-commit-id: f831ab29be9cfdc50a3034b3e270d156f37f2e6f